+++ GraphsLib/Algorithms.cs
