+++ GraphsTest/GraphsTest.csproj
@@ -51,6 +51,19 @@
     <Compile Include="Form1.Designer.cs">
       <DependentUpon>Form1.cs</DependentUpon>
     </Compile>
+    <Compile Include="GraphFormContextMenus.cs">
+      <SubType>Form</SubType>
+    </Compile>
+    <Compile Include="GraphFormMouse.cs">
+      <SubType>Form</SubType>
+    </Compile>
+    <Compile Include="GraphFormRender.cs">
+      <SubType>Form</SubType>
+    </Compile>
+    <Compile Include="GraphFormToolStrips.cs">
+      <SubType>Form</SubType>
+    </Compile>
+    <Compile Include="GraphicsExtension.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Resource1.Designer.cs">
